+++ src/components/common/Input/css/CheckBox.module.css
